Adding splash component
