+++ src/views/Lobby/index.js
@@ -1,122 +1,79 @@
-import React from 'react';
-import { Constants, Linking } from 'expo';
-import propTypes from 'prop-types';
-import { connect } from 'react-redux';
-import * as Animatable from 'react-native-animatable';
+import React, { Component } from 'react';
 import {
-  Container,
-  Content,
-  Label,
-  Button,
-  Text,
-  View,
-  List,
-  ListItem,
-  Left,
-  Footer,
-  Right,
-  Icon,
-  FooterTab,
+  Container, Content, Item, Input, Label, Button, Text
 } from 'native-base';
-import QRCode from 'react-native-qrcode';
+import { KeyboardAvoidingView } from 'react-native';
+import { connect } from 'react-redux';
+import propTypes from 'prop-types';
 import styles from './styles';
 import * as Actions from '../../actions';
 
-const Lobby = ({
-  navigation, flee, createGame, id, name, players, moderator, promote
-}) => {
-  const currentPlayers = Object.keys(players).map(k => (
-    <ListItem key={k}>
-      <Left>
-        <Animatable.Text
-          style={Constants.deviceId === k ? styles.currentPlayer : {}}
-          animation="lightSpeedIn"
-        >
-          {players[k].name}
-        </Animatable.Text>
-      </Left>
-      {moderator
-        && Constants.deviceId !== k && (
-          <Right>
-            <Icon type="Foundation" name="crown" onPress={() => promote(Constants.deviceId, k)} />
-          </Right>
-      )}
-      {players[k].moderator && (
-        <Right>
-          <Icon type="Foundation" name="crown" style={styles.activeCrown} />
-        </Right>
-      )}
-    </ListItem>
-  ));
+class Home extends Component {
+  static navigationOptions = {
+    title: 'Virtual Cards',
+  };
 
-  return (
-    <Container style={styles.content}>
-      <Content>
-        <View style={styles.qrContainer}>
-          <Label style={styles.qrLabel}>{id}</Label>
-          <QRCode value={Linking.makeUrl('/', { id })} size={200} bgColor="black" fgColor="white" />
-        </View>
-        <ListItem itemHeader style={styles.titleContainer}>
-          <Label style={styles.qrLabel}>{name}</Label>
-        </ListItem>
-        <List>{currentPlayers}</List>
-      </Content>
-      <Footer style={styles.footer}>
-        <FooterTab>
-          {(!moderator || currentPlayers.length === 1) && (
-            <Button block bordered danger style={styles.button} onPress={() => flee()}>
-              <Text>Flee</Text>
-            </Button>
-          )}
-          {moderator && (
-            <Button
-              block
-              bordered
-              warning
-              style={styles.button}
-              onPress={() => navigation.navigate('Prepare')}
-            >
-              <Text>Prepare</Text>
-            </Button>
-          )}
-          {moderator && (
-            <Button block bordered success style={styles.button} onPress={() => createGame()}>
-              <Text>Start</Text>
-            </Button>
-          )}
-        </FooterTab>
-      </Footer>
-    </Container>
-  );
-};
+  constructor() {
+    super();
+    this.state = {
+      username: '',
+    };
+  }
 
-Lobby.propTypes = {
+  onJoin = () => {
+    const { navigation, setUsername } = this.props;
+    const { username } = this.state;
+    if (username) setUsername(username);
+    navigation.navigate('Join');
+  };
+
+  createParty = async () => {
+    const { navigation, createParty, setUsername } = this.props;
+    const { username } = this.state;
+    if (username) setUsername(username);
+    await createParty();
+    navigation.navigate('Party');
+  };
+
+  onNameChange = username => this.setState({ username });
+
+  render() {
+    return (
+      <Container>
+        <KeyboardAvoidingView style={styles.keyboardAvoidingView} behavior="padding" enabled>
+          <Content contentContainerStyle={styles.content} scrollEnabled={false}>
+            <Item floatingLabel>
+              <Label>Name</Label>
+              <Input onChangeText={this.onNameChange} />
+            </Item>
+            <Button block bordered success style={styles.button} onPress={this.createParty}>
+              <Text>Create</Text>
+            </Button>
+            <Button block bordered info style={styles.button} onPress={this.onJoin}>
+              <Text>Join</Text>
+            </Button>
+          </Content>
+        </KeyboardAvoidingView>
+      </Container>
+    );
+  }
+}
+
+Home.propTypes = {
   navigation: propTypes.shape({
     navigate: propTypes.func,
-    state: propTypes.object,
   }).isRequired,
-  id: propTypes.string.isRequired,
-  name: propTypes.string.isRequired,
-  moderator: propTypes.bool.isRequired,
-  players: propTypes.shape().isRequired,
-  flee: propTypes.func.isRequired,
-  createGame: propTypes.func.isRequired,
-  promote: propTypes.func.isRequired,
+  setUsername: propTypes.func.isRequired,
 };
 
-const mapStateToProps = ({ party: { id, name, players = {} } }) => ({
-  id,
-  name,
-  players,
-  moderator: players[Constants.deviceId] ? !!players[Constants.deviceId].moderator : false,
-});
+const mapStateToProps = (state = {}) => ({ username: state.username });
+
 const mapDispatchToProps = {
-  flee: Actions.flee,
-  createGame: Actions.createGame,
-  promote: Actions.promote,
+  setUsername: Actions.setUsername,
+  createParty: Actions.createParty,
 };
 
 export default connect(
   mapStateToProps,
   mapDispatchToProps
-)(Lobby);
+)(Home);
