+++ src/views/Lobby/styles.js
@@ -1,39 +1,18 @@
 import { StyleSheet } from 'react-native';
 
 export default StyleSheet.create({
-  qrContainer: {
+  content: {
     flex: 1,
+    backgroundColor: '#fff',
     alignItems: 'center',
     justifyContent: 'center',
-    marginTop: 5,
-  },
-  currentPlayer: {
-    color: 'mediumblue',
-    fontWeight: 'bold',
-  },
-  content: {
-    backgroundColor: '#fff',
-  },
-  activeCrown: {
-    color: 'gold',
-  },
-  qrLabel: {
-    fontSize: 20,
-    marginBottom: 5,
-  },
-  titleContainer: {
-    justifyContent: 'center',
-  },
-  title: {
-    fontSize: 15,
-  },
-  footer: {
-    marginBottom: 20,
   },
   button: {
-    backgroundColor: '#fff',
     marginTop: 10,
     marginLeft: 5,
     marginRight: 5,
+  },
+  keyboardAvoidingView: {
+    flex: 1,
   },
 });
